RESTv1/requirements.txt
python-dotenv
fastapi[all]
uvicorn[standard]
deta